# target/generated-sources/openapi/src/main/java/org/openapitools/api/DeleteinternApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-14T12:32:47.611557+05:30[Asia/Kolkata]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "intern-controller", description = "the intern-controller API")
public interface DeleteinternApi {

    default DeleteinternApiDelegate getDelegate() {
        return new DeleteinternApiDelegate() {};
    }

    /**
     * DELETE /deleteintern
     *
     * @param internId  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteintern",
        tags = { "intern-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/deleteintern",
        produces = { "*/*" }
    )
    
    default ResponseEntity<String> deleteintern(
        @NotNull @Parameter(name = "internId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "internId", required = true) Integer internId
    ) {
        return getDelegate().deleteintern(internId);
    }

}