# target/generated-sources/openapi/src/main/java/org/openapitools/api/UpdateinternApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Intern;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-14T12:32:47.611557+05:30[Asia/Kolkata]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "intern-controller", description = "the intern-controller API")
public interface UpdateinternApi {

    default UpdateinternApiDelegate getDelegate() {
        return new UpdateinternApiDelegate() {};
    }

    /**
     * POST /updateintern
     *
     * @param internId  (required)
     * @param intern  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateintern",
        tags = { "intern-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/updateintern",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<String> updateintern(
        @NotNull @Parameter(name = "internId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "internId", required = true) Integer internId,
        @Parameter(name = "Intern", description = "", required = true) @Valid @RequestBody Intern intern
    ) {
        return getDelegate().updateintern(internId, intern);
    }

}